scripts/requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.29.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
python-multipart
pymysql
schedule
socks
requests-tor 
pysocks
cryptography